Add Dockerfiles for Postgres and PLV8, update README and create Makefile
This commit includes the creation of Dockerfiles for different PostgreSQL versions along with the PLV8 JavaScript extension. It also updates the README file with usage instructions and adds a Makefile for easier project management. The Dockerfiles facilitate the building and deployment of Docker images with PostgreSQL and PLV8 using various configurations.